pyproject.toml
[tool.poetry]
name = "dummyproject"
version = "0.1.0"
description = "dbt dummy project"
authors = ["Josh Taylor <joshuataylorx@gmail.com>"]
license = "Apache"
package-mode = false

[tool.poetry.dependencies]
# dbt-models requires Python 3.11
python = ">=3.11, <3.12"

# tree-sitter is tied to 0.19.0 for tree-sitter-jinja2
tree-sitter = "0.19.0"
tree-sitter-jinja2 = "0.1.0a1"

# dbt-core 1.7.11 - 2024-03-28
# https://github.com/dbt-labs/dbt-core/releases/tag/v1.7.11
dbt-core = "1.7.11"

# dbt-snowflake 1.7.2 - 2024-02-16
# https://github.com/dbt-labs/dbt-snowflake/releases/tag/v1.7.2
# Note: dbt-snowflake is a different version than dbt-core
# dbt-snowflake = "1.7.2"
dbt-postgres = "^1.7.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"